[refactor] Apply Naver Formatter
