+++ src/main/java/org/umc/travlocksserver/TravlocksServerApplication.java
@@ -6,8 +6,8 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 @SpringBootApplication
 public class TravlocksServerApplication {
 
-    public static void main(String[] args) {
-        SpringApplication.run(TravlocksServerApplication.class, args);
-    }
+	public static void main(String[] args) {
+		SpringApplication.run(TravlocksServerApplication.class, args);
+	}
 
 }
