--- src/main/java/org/umc/travlocksserver/global/apiPayload/code/BaseCode.java
@@ -1,13 +1,0 @@
-package org.umc.travlocksserver.global.apiPayload.code;
-
-import org.springframework.http.HttpStatus;
-
-public interface BaseCode {
-
-    HttpStatus getStatus();
-    String getMessage();
-
-    default String getCode() {
-        return ((Enum<?>) this).name();
-    }
-}
