--- src/main/java/org/umc/travlocksserver/global/apiPayload/exception/MemberException.java
@@ -1,11 +1,0 @@
-package org.umc.travlocksserver.global.apiPayload.exception;
-
-import org.umc.travlocksserver.global.apiPayload.code.common.CommonErrorCode;
-
-public class MemberException extends GeneralException {
-
-    // MemberErrorCode 생성하면 아래 CommonErrorCode -> MemberErrorCode로 변경
-    public MemberException(CommonErrorCode code) {
-        super(code);
-    }
-}
