+++ src/main/java/org/umc/travlocksserver/global/code/BaseCode.java
@@ -1,0 +1,14 @@
+package org.umc.travlocksserver.global.code;
+
+import org.springframework.http.HttpStatus;
+
+public interface BaseCode {
+
+	HttpStatus getStatus();
+
+	String getMessage();
+
+	default String getCode() {
+		return ((Enum<?>)this).name();
+	}
+}
